+++ src/platform_impl/linux/kms/mod.rs
@@ -3,10 +3,35 @@ use crate::dpi::{PhysicalPosition, PhysicalSize};
 pub mod event_loop;
 pub mod window;
 pub use drm::SystemError;
+use drm::{control::Device as ControlDevice, Device};
 pub use event_loop::EventLoop;
 pub use event_loop::EventLoopProxy;
 pub use event_loop::EventLoopWindowTarget;
+use std::os::unix::prelude::FromRawFd;
+use std::sync::Arc;
 pub use window::Window;
+
+#[derive(Debug, Clone)]
+/// A simple wrapper for a device node.
+pub struct Card(pub(crate) Arc<i32>);
+
+/// Implementing `AsRawFd` is a prerequisite to implementing the traits found
+/// in this crate. Here, we are just calling `as_raw_fd()` on the inner File.
+impl std::os::unix::io::AsRawFd for Card {
+    fn as_raw_fd(&self) -> std::os::unix::io::RawFd {
+        *self.0
+    }
+}
+
+impl Drop for Card {
+    fn drop(&mut self) {
+        unsafe { std::fs::File::from_raw_fd(*self.0) };
+    }
+}
+
+/// With `AsRawFd` implemented, we can now implement `drm::Device`.
+impl Device for Card {}
+impl ControlDevice for Card {}
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct DeviceId;
